Fix Q, add Q_met and test clinical case

+++ getMNL.m
@@ -4,15 +4,15 @@ function [ m, n, ell ] = getMNL(idx, x_idx_max, y_idx_max, z_idx_max)
     
     tmp_m = mod( idx, x_idx_max );
     if tmp_m == 0
-        m = x_idx_max;
+        m = int64(x_idx_max);
     else
-        m = tmp_m;
+        m = int64(tmp_m);
     end
 
     if mod( idx, x_idx_max * y_idx_max ) == 0
-        n = y_idx_max;
+        n = int64(y_idx_max);
     else
-        n = ( mod( idx, x_idx_max * y_idx_max ) - m ) / x_idx_max + 1;
+        n = int64( ( mod( idx, x_idx_max * y_idx_max ) - m ) / x_idx_max + 1 );
     end
     
     ell = int64( ( idx - m - ( n - 1 ) * x_idx_max ) / ( x_idx_max * y_idx_max ) + 1 );
